- changed manager XML parsing to use anyxml   for unknown nodes, instead of barfing on the message - Added port leaf-list CLI parm to netconfd - Fixed bug in agent message processing that killed   the session description but not the active_fd_set in agt_ncxserver - Fixed bugs in val.c; NCX_BT_BITS and NCX_BT_KEYREF not supported   in some utility functions
git-svn-id: https://yuma.svn.sourceforge.net/svnroot/yuma/trunk@121 1ffd1d58-2efc-45df-af60-190a33ec9863

+++ netconf/src/agt/agt_ses.h
@@ -99,7 +99,7 @@ extern boolean
     agt_ses_process_first_ready (void);
 
 extern boolean
-    agt_ses_ssh_port_allowed (uint32 port);
+    agt_ses_ssh_port_allowed (uint16 port);
 
 extern void
     agt_ses_fill_writeset (fd_set *fdset,
